requirements.txt
python-docx>=1.1.0
jira>=3.8.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
playwright>=1.49.0
requests>=2.32.0
pydantic>=2.7.0
langchain>=0.3.0
openai>=1.55.0
python-dotenv>=1.0.1
uvloop>=0.19.0
